sdk-py/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "openmemory-py"
version = "1.0.0"
description = "OpenMemory Python SDK provides a local-first long-term memory engine for AI agents and LLM applications. Features include semantic search, multi-sector memory, temporal fact storage, automatic decay, and explainable recall paths. Works fully offline or with the OpenMemory backend."
readme = "README.md"
keywords = "ai llm memory long-term-memory agent agents semantic-memory episodic-memory vector-search persistent-memory llm-memory cognitive-memory temporal-graph mcp claude openai python sdk"
authors = [{ name = "OpenMemory Team" }]
license = { text = "Apache-2.0" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "google-generativeai>=0.3.0",
    "boto3>=1.34.0",
]

[project.urls]
"Homepage" = "https://github.com/cavira/OpenMemory"
"Bug Tracker" = "https://github.com/cavira/OpenMemory/issues"

[tool.setuptools.packages.find]
where = ["src"]
